+++ extensions/i18n/assets/BaseLocalized.ts
@@ -1,5 +1,5 @@
 import { Component } from 'cc';
 
 export abstract class BaseLocalized extends Component {
-    public updateRenderer() {}
+    public localize() {}
 }