+++ extensions/i18n/assets/LocalizedSprite.ts
@@ -31,7 +31,7 @@ export class LocalizedSprite extends BaseLocalized {
                 // i18n.init('en');
                 return;
             }
-            this.updateRenderer();
+            this.localize();
         }
     }
 
@@ -45,15 +45,20 @@ export class LocalizedSprite extends BaseLocalized {
         }
     }
 
-    async updateRenderer() {
+    async localize() {
         // @ts-ignore
         if (CC_EDITOR) {
             this.removeListener();
-            i18n._language;
-            let uuid = await Editor.Message.request('asset-db', 'query-uuid', eval('`' + this.spriteUrl + '`'));
+            let url = eval('`' + this.spriteUrl + '`');
+            let uuid = await Editor.Message.request('asset-db', 'query-uuid', url);
+            if(!uuid) {
+                const defaultPath = this.spriteUrl.replace('${i18n._language}', i18n._default);
+                uuid = await Editor.Message.request('asset-db', 'query-uuid', defaultPath);
+                console.warn(`${this.node.name} use default language ${i18n._default} instead of ${i18n._language}`);
+            }
             await Editor.Message.request('scene', 'set-property', {
                 uuid: this.node.uuid,
-                path: `__comps__.${this.getComponents(Component).findIndex(val=>val==this.sprite)}.spriteFrame`,
+                path: `__comps__.${this.getComponents(Component).findIndex((val) => val == this.sprite)}.spriteFrame`,
                 dump: {
                     type: 'cc.SpriteFrame',
                     value: {
@@ -68,7 +73,7 @@ export class LocalizedSprite extends BaseLocalized {
                 .catch((err) => {
                     // 一段時間後retry
                     setTimeout(() => {
-                        this.updateRenderer();
+                        this.localize();
                     }, this.retryInterval);
                 });
         }
@@ -94,7 +99,7 @@ export class LocalizedSprite extends BaseLocalized {
             // 1. 將多國語系前面刪除 xxx/yyy.png@zzzzz
             // 2. 將.png後移除 xxx/yyy
             // 3. 補上spriteFrame檔 xxx/yyy/spriteFrame
-            let path = this.spriteUrl.split('${i18n._language}/')[1].split('.png')[0] + '/spriteFrame';
+            let path = this.spriteUrl.split('${i18n._language}/')[1].replace(/\.\w+@\w+/, '') + '/spriteFrame';
             bundle.load(path, SpriteFrame, (err, spriteFrame) => {
                 if (err) {
                     return reject(err);
@@ -112,12 +117,22 @@ export class LocalizedSprite extends BaseLocalized {
     }
 
     private processPath() {
-        //db://assets/resources/language/en/xxxxxx
-        // console.log(this.spriteUrl);
-        // const regex = 'language/[a-z]+/\\w+';
-        // this.spriteUrl = this.spriteUrl.slice(this.spriteUrl.search(regex));
-        const splitArray = this.spriteUrl.split('/');
-        this.spriteUrl = this.spriteUrl.replace(splitArray[splitArray.indexOf('language') + 1], '${i18n._language}'); //.split('@')[0];
+        // 調整路徑處理方式，不在綁定在language下，但各語系仍需要在同個資料夾下且都打包bundle，且不支援複數bundle
+        let done = false;
+        JSON.stringify(window.languages, (key, value) => {
+            if (done) {
+                return undefined;
+            } else {
+                if (key != '' && window.languages[key]) {
+                    const reg = new RegExp(`\/${key}\/`);
+                    if (this.spriteUrl.match(reg)) {
+                        this.spriteUrl = this.spriteUrl.replace(reg, '/${i18n._language}/');
+                        done = true;
+                    }
+                }
+                return value;
+            }
+        });
     }
 
     public onDisable() {