+++ extensions/i18n/i18n/en.js
@@ -5,5 +5,8 @@ module.exports = {
     'menu': 'Open i18n Panel',
     'title': 'i18n Panel',
     'clear': 'clear Reference',
-    'magic': 'magic revert'
+    'magic': 'magic revert',
+    'buildPanel': {
+        'title' : 'Default Language'
+    }
 };