+++ extensions/i18n/i18n/zh.js
@@ -5,5 +5,8 @@ module.exports = {
     'menu':  '開啟多國語系面板',
     'title': '多國語系面板',
     'clear': '清除資源關聯',
-    'magic': '神奇復原'
+    'magic': '神奇復原',
+    'buildPanel': {
+        'title' : '預設語系'
+    }
 };