+++ extensions/i18n/src/hooks.ts
@@ -3,11 +3,16 @@ import fs from 'fs';
 import { join } from 'path';
 import { IBuildResult, IBuildTaskOption } from '../@types/packages/builder/@types';
 
+const autoAtlasContent = `{
+    "__type__": "cc.SpriteAtlas"
+}`;
+
 interface IOptions {
-    webTestOption: boolean;
+    defaultLang: string;
 }
 
 const PACKAGE_NAME = 'i18n';
+const languagePath = `db://assets/art/language`;
 
 interface ITaskOptions extends IBuildTaskOption {
     packages: {
@@ -19,42 +24,43 @@ function log(...arg: any[]) {
     return console.log(`[${PACKAGE_NAME}] `, ...arg);
 }
 
-let allAssets = [];
-
 export const throwError = true;
 
 export async function load() {
-    console.log(`[${PACKAGE_NAME}] Load cocos plugin example in builder.`);
-    allAssets = await Editor.Message.request('asset-db', 'query-assets');
+    log(`Load cocos plugin example in builder.`);
 }
 
+//#region onBeforeBuild
 export async function onBeforeBuild(options: ITaskOptions) {
-    // Todo some thing
-    // if (options.packages[PACKAGE_NAME].webTestOption) {
+    log(`onBeforeBuild`);
     await removeReference();
-    // }
+    await fixSpineImageMeta();
+    await fillResources(options.packages[PACKAGE_NAME].defaultLang);
 }
 
 async function removeReference() {
     return new Promise<void>(async (resolve, reject) => {
+        let metaData = await fs.readFileSync(
+            join(Editor.Project.path, 'extensions/i18n/assets/LocalizedSprite.ts.meta'),
+            'utf8'
+        );
+        let localizedSpriteUuid = Editor.Utils.UUID.compressUUID(JSON.parse(metaData).uuid, false);
+
+        metaData = await fs.readFileSync(
+            join(Editor.Project.path, 'extensions/i18n/assets/LocalizedSkeleton.ts.meta'),
+            'utf8'
+        );
+        let localizedSkeletonUuid = Editor.Utils.UUID.compressUUID(JSON.parse(metaData).uuid, false);
+
         let assets: AssetInfo[] = [];
         await Editor.Message.request('asset-db', 'query-assets', {
             pattern: 'db://assets/**'
         }).then((infos) => Array.prototype.push.apply(assets, infos.filter(filterFn)));
 
-        // await Editor.Message.request('asset-db', 'query-assets', {
-        //     pattern: 'db://assets/scenes/**'
-        // }).then((infos) => Array.prototype.push.apply(assets, infos.filter(filterFn)));
-        // assets.forEach((data) => console.log(`asset : ${data.name} , ${data.type} , ${data.url}`));
-        let path = join(Editor.Project.path, 'extensions/i18n/assets/LocalizedSprite.ts.meta');
-        let rawData = fs.readFileSync(path, 'utf8');
-        let data = JSON.parse(rawData);
-        let uuid = Editor.Utils.UUID.compressUUID(data.uuid, false);
-        // console.log(`uuid = ${uuid}, uuid compare : ${uuid == "c05c25C4xNAupWYMB93tIzp"}`);
         for (let i = 0; i < assets.length; i++) {
-            await setNullToReference(uuid, assets[i]);
+            await setNullToReference(assets[i], localizedSpriteUuid, localizedSkeletonUuid);
         }
-        console.log(`${PACKAGE_NAME} removeReference end`);
+        log(`removeReference end`);
         resolve();
     });
 }
@@ -63,50 +69,165 @@ function filterFn(info: AssetInfo) {
     return info.type == 'cc.Prefab' || info.type == 'cc.SceneAsset';
 }
 
-async function setNullToReference(uuid: string, info: AssetInfo) {
-    //db://assets/prefabs/Button.prefab
-    let path = join(Editor.Project.path, info.url.slice(5));
+async function setNullToReference(info: AssetInfo, localizedSpriteUuid: string, localizedSkeletonUuid: string) {
+    let path = join(Editor.Project.path, info.url.replace('db://', ''));
     let data = fs.readFileSync(path, 'utf8');
-    if (data.includes(uuid)) {
-        let json = JSON.parse(data);
-        // get all LocalizedSprite
-        const target = json.filter((element: { __type__: string }) => element.__type__ == uuid);
+    let needToSave = false;
+    if (data.includes(localizedSpriteUuid)) {
+        data = removeSpriteReference(data, localizedSpriteUuid);
+        needToSave = true;
+    }
 
-        for (let i = 0; i < target.length; i++) {
-            const targetNodeID = target[i].node.__id__;
-            json.find(
-                (element: { __type__: string; node: { __id__: any } }) =>
-                    element.__type__ == 'cc.Sprite' && element.node.__id__ == targetNodeID
-            )._spriteFrame = null;
-        }
-        data = JSON.stringify(json);
+    if (data.includes(localizedSkeletonUuid)) {
+        data = removeSkeletonReference(data, localizedSkeletonUuid);
+        needToSave = true;
+    }
 
+    if (needToSave) {
         fs.writeFileSync(path, data, 'utf8');
         await Editor.Message.request('asset-db', 'reimport-asset', info.url);
     }
 }
 
-export async function onBeforeBuildAssets(options: ITaskOptions, result: IBuildResult) {
-    // Todo some thing
-    log(`${PACKAGE_NAME}.webTestOption`, 'onBeforeBuildAssets');
-}
+function removeSpriteReference(data: string, uuid: string) {
+    let json = JSON.parse(data);
+    // get all LocalizedSprite
+    const target = json.filter((element: { __type__: string }) => element.__type__ == uuid);
 
-export async function onBeforeCompressSettings(options: ITaskOptions, result: IBuildResult) {
-    const pkgOptions = options.packages[PACKAGE_NAME];
-    if (pkgOptions.webTestOption) {
-        console.debug('webTestOption', true);
+    for (let i = 0; i < target.length; i++) {
+        const targetNodeID = target[i].node?.__id__;
+        if (targetNodeID) {
+            json.find(
+                (element: { __type__: string; node: { __id__: any } }) =>
+                    element.__type__ == 'cc.Sprite' && element.node.__id__ == targetNodeID
+            )._spriteFrame = null;
+        }
     }
-    // Todo some thing
-    console.debug('get settings test', result);
+    return JSON.stringify(json);
 }
 
-export async function onAfterCompressSettings(options: ITaskOptions, result: IBuildResult) {
-    // Todo some thing
-    console.log('webTestOption', 'onAfterCompressSettings');
+function removeSkeletonReference(data: string, uuid: string) {
+    let json = JSON.parse(data);
+    // get all LocalizedSkeleton
+    while (true) {
+        let target = json.find(
+            (element: { __type__: string; _skeletonData: string }) =>
+                element.__type__ == uuid && element._skeletonData != null
+        );
+        if (target) {
+            target._skeletonData = null;
+        } else {
+            break;
+        }
+    }
+    return JSON.stringify(json);
 }
 
-export async function onAfterBuild(options: ITaskOptions, result: IBuildResult) {}
+async function fixSpineImageMeta() {
+    let imageInfos = await Editor.Message.request('asset-db', 'query-assets', {
+        pattern: 'db://assets/**',
+        ccType: 'cc.ImageAsset'
+    });
+
+    let spinePaths = Array.from(
+        await Editor.Message.request('asset-db', 'query-assets', {
+            pattern: 'db://assets/**',
+            ccType: 'sp.SkeletonData'
+        }),
+        (info) => info.path
+    );
+
+    for (let info of imageInfos) {
+        if (checkSpinePath(info.path, spinePaths) != '') {
+            let metaFile = `${info.file}.meta`;
+            let metaData = fs.readFileSync(metaFile, 'utf8').replace(/"packable"\s*:\s*true\s*/, `"packable": false`);
+            fs.writeFileSync(metaFile, metaData, 'utf8');
+            await Editor.Message.request('asset-db', 'reimport-asset', info.url);
+        }
+    }
+}
+
+async function fillResources(defaultLang: string) {
+    const sourcePath = `${languagePath}/${defaultLang}/**`;
+    let sources = await Editor.Message.request('asset-db', 'query-assets', {
+        pattern: sourcePath
+    });
+
+    //${path}.skel ${path}.atlas ${path}.png ${path}2.png ....
+    let spinePaths = Array.from(
+        await Editor.Message.request('asset-db', 'query-assets', {
+            pattern: sourcePath,
+            ccType: 'sp.SkeletonData'
+        }),
+        (info) => info.path
+    );
+
+    const languages = Array.from(
+        await Editor.Message.request('asset-db', 'query-assets', {
+            pattern: `${languagePath}/**`,
+            isBundle: true
+        }),
+        (info) => info.name
+    );
+
+    for (const lang of languages) {
+        await fillAutoAtlas(lang);
+        if (lang != defaultLang) {
+            await fillTargetResource(sources, spinePaths, defaultLang, lang);
+        }
+    }
+}
+
+async function fillTargetResource(sources: AssetInfo[], spinePaths: string[], defaultLang: string, targetLang: string) {
+    // xxx.skel xxx.atlas xxx.png xxx2.png ...
+    for (let i = 0; i < sources.length; i++) {
+        const source = sources[i].source;
+        //ignore auto-atlas
+        if (source.endsWith('.pac')) {
+            continue;
+        }
+        const target = source.replace(`/${defaultLang}/`, `/${targetLang}/`);
+        if (await isExist(target)) {
+            let spinePath = checkSpinePath(source, spinePaths);
+            if (spinePath != '') {
+                const lastSpine = sources.filter((info) => info.source.startsWith(spinePath)).slice(-1)[0];
+                i = sources.indexOf(lastSpine);
+            }
+        } else {
+            await Editor.Message.request('asset-db', 'copy-asset', source, target);
+        }
+    }
+}
+
+function checkSpinePath(fullPath: string, spinePaths: string[]) {
+    for (let path of spinePaths) {
+        if (fullPath.startsWith(path)) {
+            return path;
+        }
+    }
+    return '';
+}
+
+async function isExist(path: string) {
+    return (await Editor.Message.request('asset-db', 'query-asset-info', path)) != null;
+}
+
+async function fillAutoAtlas(lang: string) {
+    const autoAtlas = await Editor.Message.request('asset-db', 'query-assets', {
+        pattern: `${languagePath}/${lang}/**`,
+        ccType: 'cc.SpriteAtlas'
+    });
+    if (autoAtlas.length == 0) {
+        await Editor.Message.request(
+            'asset-db',
+            'create-asset',
+            `db://assets/art/language/${lang}/auto-atlas.pac`,
+            autoAtlasContent
+        );
+    }
+}
+//#endregion
 
 export function unload() {
-    console.log(`[${PACKAGE_NAME}] Unload cocos plugin example in builder.`);
+    log(`Unload cocos plugin example in builder.`);
 }